Added the ability to process files in Config/Input
This allows users to edit the Config/Input/Aircrafts bindings, such as the base and common bindings, which is especially useful to set up general bindings for the lower fidelity modules such as in FC3. Only tested with the Aircrafts subfolder, as apparently it has a differing prefix than the UiLayer despite being in the same directory.