Updated data.lua to DCS World Version 2.7.9.18080 and added the ability to process files in Config/Input
This allows users to edit the Config/Input/Aircrafts bindings, such as the base and common bindings, which is especially useful to set up general bindings for the lower fidelity modules such as in FC3. Only tested with the Aircrafts subfolder, as apparently it has a differing prefix than the UiLayer despite being in the same directory.
Also included are DCS Update 2.7.8.18080 changes to data, namely the addition of the SideWinder Force Feedback 2 Joystick in the default assignments.